@@ -1,17 +1,7 @@
-import { pascalCaseFilename } from "js/utils"
+import Configs from "./configs"
+import Topbar from "./topbar"
 
-const request = require.context(".", true, /\.jsx?$/)
-
-request.keys().forEach( function( key ){
-  if( key === "./index.js" ) {
-    return
-  }
-
-  // if( key.slice(2).indexOf("/") > -1) {
-  //   // skip files in subdirs
-  //   return
-  // }
-
-  let mod = request(key)
-  module.exports[pascalCaseFilename(key)] = mod.default ? mod.default : mod
-})
+export default {
+  Configs,
+  Topbar
+}
